updated AddToMyLists component
