+++ frontend/src/App.tsx
@@ -59,12 +59,12 @@ export default function App() {
             {user && <Header user={user}/>}
             <Routes>
                 <Route path={"/"} element={user === null && <LoginPage user={user}/>}/>
-                <Route path={"/sign_up"} element={<SignUpPage fetchUser={fetchUser}/>}></Route>
+                <Route path={"/sign_up"} element={user && <SignUpPage fetchUser={fetchUser} user={user}/>}></Route>
                 <Route element={<ProtectedRoutes user={user}/>}>
                     <Route path={"/home"} element={user && <Homepage user={user} boulders={boulders} fetchData={fetchData}/>}/>
                     <Route path={"/boulder/:id"} element={user && <BoulderDetails boulders={boulders} fetchData={fetchData} user={user} fetchUser={fetchUser}/>}/>
                     <Route path={"/profile/:tabName?"} element={user && <ProfilePage user={user}/>}/>
-                    <Route path={"/editProfile"} element={<EditProfile/>}/>
+                    <Route path={"/editProfile"} element={user && <EditProfile fetchUser={fetchUser} user={user}/>}/>
                 </Route>
             </Routes>
         </div>
