+++ frontend/src/pages/EditProfile.tsx
@@ -1,7 +1,15 @@
-export default function EditProfile(){
+import FormComponent from "../components/FormComponentProps";
+import {User} from "../types/User.ts";
+
+type EditProfileProps={
+    fetchUser: ()=>void;
+    user: User;
+}
+export default function EditProfile(props: Readonly<EditProfileProps>){
     return(
         <div>
-            Edit your profile here
+            <p>Edit your profile here</p>
+            <FormComponent fetchUser={props.fetchUser} user={props.user} path={"/profile"}/>
         </div>
     )
 }
